Restructuring of folders
