+++ gulpfile.js
@@ -10,7 +10,7 @@ var gulp = require('gulp'),
 
 // Define paths
 var paths = {
-  indexJS: ['./src/js/index.js'],
+  indexJS: ['./src/js/pages/index.js'],
   js: ['src/js/*.js']
 };
