--- src/js/Style.js
@@ -1,29 +1,0 @@
-var style = {
-  content: {
-    overflowY: 'scroll',
-    width: '100%',
-    height: '250px'
-  },
-    dropzone: {
-      width: '100%',
-      color: '#212121',
-      height: '200px',
-      border: '2px dashed #727272',
-      borderRadius: '5px',
-      fontSize: '25px'
-    },
-    dropzoneActive: {
-      border: '2px solid #727272',
-      backgroundColor: '#f5f5f5'
-    },
-    dropzoneText: {
-      lineHeight: '200px',
-      verticalAlign: 'middle',
-      width: '100%'
-    },
-  header: {
-    textAlign: 'center'
-  }
-};
-
-module.exports = style;
