+++ src/js/actions/CoreActionConsts.js
