+++ src/js/components/BookTitle.js
@@ -7,7 +7,7 @@ const React = require('react');
 const BookTitle = React.createClass({
   render: function() {
     return (
-   <h4>{this.props.title} </h4>
+   <h4> {this.props.title} </h4>
  );
   }
 });
