+++ src/js/components/CheckModule.js
