+++ src/js/components/CheckingScreen.js
