+++ src/js/components/Pane.js
@@ -9,9 +9,9 @@ const Pane = React.createClass({
   render: function() {
     return (
       <Col md={4} sm={4} xs={12}>
-          <h3 style={style.header}> {this.props.title} </h3>
+        <h3 style={style.header}> {this.props.title} </h3>
           <Well style={style.content}>
-          <div>{this.props.content}</div>
+            <div>{this.props.content}</div>
           </Well>
       </Col>
     );
