+++ src/js/components/Verse.js
